Remove all u8s

+++ benches/similarities.rs
@@ -39,7 +39,7 @@ fn similarities(c: &mut Criterion) {
         let shape_rotated: StrongPoints<shapes::Vertex> = StrongPoints::new(shapes::similarity::unit_sphere_normalize(rotation * shape_coordinates));
         let bijection: Bijection<shapes::Vertex, shapes::Column> = {
             let mut p = permutation::Permutation::random(shape.size());
-            p.sigma.push(p.set_size() as u8);
+            p.sigma.push(p.set_size());
             strong::bijection::Bijection::new(p)
         };
         let cloud = shape_rotated.apply_bijection(&bijection);
